Incremental work on buildUpdateRequest(). Small bugfixes.

+++ lib/html/header.go
@@ -25,8 +25,8 @@ func writeHeader(writer io.Writer) {
 	sysCpuTime := time.Duration(rusage.Stime.Sec)*time.Second +
 		time.Duration(rusage.Stime.Usec)*time.Microsecond
 	cpuTime := rusage.Utime.Sec + rusage.Stime.Sec
-	fmt.Fprintf(writer, "CPU Time: %d%% (User: %s Sys: %s)<br>\n",
-		cpuTime*100/int64(uptime.Seconds()), userCpuTime, sysCpuTime)
+	fmt.Fprintf(writer, "CPU Time: %.1f%% (User: %s Sys: %s)<br>\n",
+		float64(cpuTime*100)/float64(uptime.Seconds()), userCpuTime, sysCpuTime)
 	var memStatsBeforeGC, memStatsAfterGC runtime.MemStats
 	runtime.ReadMemStats(&memStatsBeforeGC)
 	runtime.GC()
